\name{gcd}
\alias{gcd}
\title{
Greatest Common Divisor
}
\description{
Computes the GCD of two numbers.}

\usage{
gcd(v, t)
}

\arguments{
  \item{v}{first input number.}
  \item{t}{second input number.}
}

\value{
Computes the GCD of the two given numbers
}

\references{
https://rosettacode.org/wiki/Greatest_common_divisor#R
}

\examples{
gcd(6,12)
}

\keyword{gcd}