\name{vector_gcd}
\alias{vector_gcd}
\title{
Vector Greatest Common Divisor
}
\description{
Computes the GCD of a vector.
}
\usage{
vector_gcd(vec)
}

\arguments{
  \item{vec}{a vector, for which we want to compute the GCD.}
}

\value{
The GCD of the elements in the given vector.
}


\examples{
input = c(3,6,9)
vector_gcd(input)
}

\keyword{gcd}
\keyword{vector}
